requirements.txt
playwright>=1.40.0
cryptography>=41.0.0
Pillow>=10.0.0
PyYAML>=6.0.1
requests>=2.31.0
